dropped database and created a new one changing satisfaction and revisit data types to float

+++ db/migrate/20191214154956_create_reviews.rb
@@ -2,8 +2,8 @@ class CreateReviews < ActiveRecord::Migration[6.0]
   def change
     create_table :reviews do |t|
       t.text :occasion
-      t.integer :satisfaction
-      t.integer :revisit
+      t.float :satisfaction
+      t.float :revisit
       t.text :suggest
       t.text :living
       t.text :stay_length
